requirements.txt
Django==1.5.4
Fabric==1.8.0
Pillow==1.7.8
Scrapy==0.20.0
South==0.8.2
cssselect==0.8
django-arrayfields==0.1.0
django-debug-toolbar==0.9.4
django-extensions==0.8
django-grappelli==2.4.6
django-modeltranslation==0.4.1
django-nose==1.2
django-storages==1.1.8
django-tastypie==0.10.0
-e git://github.com/coagulant/django-whatever.git@62d5c8de195de88a2e4b6ce38a3698aaa24a3a62#egg=django_whatever-dev
gevent==0.13.8
grappelli-modeltranslation==0.1.2
greenlet==0.4.1 # required by gevent
lxml==3.1.2
mimeparse==0.1.3 # required by django-tastypie
mock==1.0.1
model_mommy==1.2
nose==1.3.0
paramiko==1.11.1 # required by Fabric
psycopg2==2.5.1
pycrypto==2.6 # required by paramiko
python-dateutil==2.1 # required by django-tastypie
python-memcached==1.53
requests==1.2.0
six==1.4.1 # required by python-dateutil
-e git://github.com/ostrovok-team/xact.git@65e19094eea1d281fdd03b0ad413e6b3bf6b4593#egg=xact-dev
